requirements.txt
aiogram==2.7
requests==2.28.0
python-dotenv==0.20.0